# swagger.yaml
openapi: 3.0.0
info:
  description: "This is the swagger documentation and API for the GeoCode project implemented by Peak Performers for the client [5DT](https://5dt.com/)."
  version: "1.0.0"
  title: "Swagger GeoCode"
  termsOfService: "https://swagger.io/terms/"
  contact:
    email: "peakperformers@geocodeapp.tech"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: 'http://geocodeapp.tech:8080/api'
...
paths:
...
  /GeoCode/createGeoCode:
    post:
      tags:
        - "GeoCode"
      summary: "Creates a new GeoCode"
      description: "Create GeoCode"
      operationId: "createGeoCode"
      requestBody:
        required: true
        description: "Request to create a new GeoCode"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGeoCodeRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateGeoCodeRequest'
      responses:
        '200':
          description: "Create GeoCode Response"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGeoCodeResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/CreateGeoCodeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getCollectables:
    post:
      tags:
        - "GeoCode"
      summary: "Get all Collectables for a certain GeoCode"
      description: "Get a GeoCode's Collectables"
      operationId: "getGeoCodeCollectables"
      requestBody:
        required: true
        description: "Request to get a GeoCode's Collectables"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCollectablesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetCollectablesRequest'
      responses:
        '200':
          description: "Returned the GeoCode's Collectables successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetCollectablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getGeoCodesByDifficulty:
    post:
      tags:
        - "GeoCode"
      summary: "Get all all the stored GeoCodes by the specified difficulty"
      description: "Get all GeoCodes by difficulty"
      operationId: "getGeoCodesByDifficulty"
      requestBody:
        required: true
        description: "Request to get all the GeoCodes by the specified difficulty"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetGeoCodesByDifficultyRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetGeoCodesByDifficultyRequest'
      responses:
        '200':
          description: "Returned all the GeoCode's with the specified difficulty"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGeoCodesByDifficultyResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetGeoCodesByDifficultyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getGeoCodes:
    get:
      tags:
        - "GeoCode"
      summary: "Get all the GeoCodes on the platform"
      description: "Get all the GeoCodes that are stored on the platform"
      operationId: "getGeoCodes"
      responses:
        '200':
          description: "Returned all the GeoCodes successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGeoCodesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetGeoCodesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getHints:
    post:
      tags:
        - "GeoCode"
      summary: "Get the hints for the specified GeoCode"
      description: "Get the hints for the specified GeoCode to help locate it"
      operationId: "getHints"
      requestBody:
        required: true
        description: "Request to get the hints from the specified GeoCode"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetHintsRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetHintsRequest'
      responses:
        '200':
          description: "Returned the hints for the GeoCode successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHintsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetHintsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getGeoCodeByQRCode:
    post:
      tags:
        - "GeoCode"
      summary: "Get the GeoCode associated with the given QR Code"
      description: "Get the GeoCode associated with the given QR Code"
      operationId: "getGeoCodeByQRCode"
      requestBody:
        required: true
        description: "Request to get a GeoCode's associated with the given QR Code"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetGeoCodeByQRCodeRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetGeoCodeByQRCodeRequest'
      responses:
        '200':
          description: "Returned the associated GeoCode successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGeoCodeByQRCodeResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetGeoCodeByQRCodeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getGeoCodeByLocation:
    post:
      tags:
        - "GeoCode"
      summary: "Get the GeoCode at or near the given location"
      description: "Get the GeoCode at or near the given location"
      operationId: "getGeoCodeByLocation"
      requestBody:
        required: true
        description: "Request to get a GeoCode at or near the given location"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetGeoCodeByLocationRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetGeoCodeByLocationRequest'
      responses:
        '200':
          description: "Return the GeoCodec at or near the given location"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGeoCodeByLocationResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetGeoCodeByLocationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/getTrackables:
    post:
      tags:
        - "GeoCode"
      summary: "Get the Trackable for a certain GeoCode"
      description: "Get a GeoCode's Trackable"
      operationId: "getTrackables"
      requestBody:
        required: true
        description: "Request to get a GeoCode's Trackable"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTrackablesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/GetTrackablesRequest'
      responses:
        '200':
          description: "Returned the GeoCode's Trackable successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTrackablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetTrackablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/swapCollectables:
    post:
      tags:
        - "GeoCode"
      summary: "Swap a specific GeoCode's Collectable"
      description: "Swap a specific GeoCode's Collectable"
      operationId: "swapCollectables"
      requestBody:
        required: true
        description: "Request to swap a GeoCode's Collectables"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapCollectablesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/SwapCollectablesRequest'
      responses:
        '200':
          description: "Swapped the GeoCode's Collectable successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapCollectablesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SwapCollectablesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
  /GeoCode/updateAvailability:
    post:
      tags:
        - "GeoCode"
      summary: "Update the availability for a certain GeoCode"
      description: "Update the availability for a certain GeoCode"
      operationId: "updateAvailability"
      requestBody:
        required: true
        description: "Request to update a GeoCode's availability"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAvailabilityRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/UpdateAvailabilityRequest'
      responses:
        '200':
          description: "Updated the GeoCode's availability successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAvailabilityResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/UpdateAvailabilityResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: []
...
components:
  responses:
    UnauthorizedError:
      description: Invalid JWT token
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Difficulty:
      type: string
      description: "The GeoCode's real world locating difficulty"
      enum:
        - "EASY"
        - "MEDIUM"
        - "DIFFICULTY"
        - "INSANE"
    Rarity:
      type: string
      enum:
        - "COMMON"
        - "UNCOMMON"
        - "RARE"
        - "EPIC"
        - "LEGENDARY"
    GeoCode:
      type: "object"
      required:
        - id
        - difficulty
        - description
        - hints
        - qrCode
        - location
        - available
      properties:
        id:
          type: string
          format: uuid
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        available:
          type: boolean
        description:
          type: string
        hints:
          type: array
          items:
            type: string
        collectables:
          $ref: '#/components/schemas/Collectable'
        trackables:
          type: string
        qrCode:
          type: string
        location:
          type: string
      xml:
        name: "GeoCode"
    CreateGeoCodeRequest:
      type: "object"
      required:
        - id
        - description
        - location
        - difficulty
        - hints
        - available
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        available:
          type: boolean
      xml:
        name: "CreateGeoCodeRequest"
    CreateGeoCodeResponse:
      type: "object"
      properties:
        geoCode:
          $ref: '#/components/schemas/GeoCode'
      xml:
        name: "CreateGeoCodeResponse"
    GetGeoCodesByDifficultyRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "getGeoCodesByDifficultyRequest"
    GetGeoCodesByDifficultyResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "getGeoCodesByDifficultyResponse"
    GetGeoCodesResponse:
      type: "object"
      required:
        - geocodes
      properties:
        geocodes:
          type: array
          items:
            $ref: '#/components/schemas/GeoCode'
      xml:
        name: "getGeoCodesByDifficultyResponse"
    GetHintsRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "getGeoCodesByDifficultyRequest"
    GetHintsResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "getGeoCodesByDifficultyResponse"
    GetGeoCodeByQRCodeRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "getGeoCodeByQRCodeRequest"
    GetGeoCodeByQRCodeResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "getGeoCodeByQRCodeResponse"
    GetGeoCodeByLocationRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "getGeoCodeByLocationRequest"
    GetGeoCodeByLocationResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "getGeoCodeByLocationResponse"
    GetTrackablesRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "getTrackablesRequest"
    GetTrackablesResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "getTrackablesResponse"
    SwapCollectablesRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "swapCollectablesRequest"
    SwapCollectablesResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "swapCollectablesResponse"
    UpdateAvailabilityRequest:
      type: "object"
      required:
        - description
        - location
        - difficulty
        - hints
      properties:
        description:
          type: string
        location:
          type: string
        hints:
          type: array
          items:
            type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
      xml:
        name: "updateAvailabilityRequest"
    UpdateAvailabilityResponse:
      type: "object"
      properties:
        id:
          type: string
          format: uuid
        qrCode:
          type: string
      xml:
        name: "updateAvailabilityResponse"
    GetCollectablesRequest:
      type: "object"
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
      xml:
        name: "getCollectablesRequest"
    CollectableType:
      type: "object"
      required:
        - id
        - name
        - image
        - rarity
        - set
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image:
          type: string
        rarity:
          $ref: "#/components/schemas/Rarity"
        set:
          $ref: "#/components/schemas/CollectableSet"
      xml:
        name: "CollectableType"
    CollectableSet:
      type: "object"
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      xml:
        name: "CollectableSet"
    Collectable:
      type: "object"
      required:
        - id
        - type
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/CollectableType"
      xml:
        name: "Collectable"
...
    GetCollectablesResponse:
      type: "object"
      required:
        - collectables
      properties:
        collectables:
          type: array
          items:
            $ref: '#/components/schemas/Collectable'
      xml:
        name: "CollectableResponse"
...